Added tests for database-stocks, refactor tests
